backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
motor==3.4.0
beanie==1.25.0
pymongo==4.7.2
faiss-cpu==1.8.0
numpy==1.26.4
Pillow==10.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
python-dotenv==1.0.1
pydantic==2.7.1
# GPU / AI (optional - comment out for CPU-only demo):
# facenet-pytorch==2.6.0
# torch==2.3.0
# torchvision==0.18.0
# Optional - CCTV video feature:
# opencv-python-headless==4.9.0.80
